Organize components as decided by Robbert & Kees
